// plugins_/qms3_form_brick/admin/blocks/src/types/StructureRow.ts
export interface StructureRow {
    block: string;
    label: string;
    type: string;
    name: string;
    default: string;
    prepend: string;
    append: string;
    header_notice: string;
    body_notice: string;
    options: string;
    placeholder: string;
    for_bcc: boolean;
    thanks_to: boolean;
    required: boolean;
    attributes: string;
}